@@ -1,5 +1,5 @@
 {
-  "name": "stock-trading",
+  "name": "weather-expert",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
@@ -11,23 +11,34 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "^14.0.0",
-    "@angular/cdk": "^14.0.5",
-    "@angular/common": "^14.0.0",
-    "@angular/compiler": "^14.0.0",
-    "@angular/core": "^14.0.0",
-    "@angular/forms": "^14.0.0",
-    "@angular/material": "^14.0.5",
-    "@angular/platform-browser": "^14.0.0",
-    "@angular/platform-browser-dynamic": "^14.0.0",
-    "@angular/router": "^14.0.0",
+    "@angular/common": "^14.2.3",
+    "@angular/compiler": "^14.2.3",
+    "@angular/core": "^14.2.3",
+    "@angular/forms": "^14.2.3",
+    "@angular/platform-browser": "^14.2.3",
+    "@angular/platform-browser-dynamic": "^14.2.3",
+    "@angular/pwa": "^0.6.8",
+    "@angular/router": "^14.2.3",
+    "@angular/service-worker": "14.2.3",
+    "@ngrx/effects": "^14.3.1",
+    "@ngrx/entity": "^14.3.1",
+    "@ngrx/store": "^14.3.1",
+    "@ngrx/store-devtools": "^14.3.1",
+    "body-parser": "^1.19.0",
+    "bootstrap": "^3.4.1",
+    "core-js": "^2.6.9",
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
+    "web-push": "^3.3.5",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^14.0.1",
     "@angular/cli": "~14.0.1",
     "@angular/compiler-cli": "^14.0.0",
+    "@ngrx/schematics": "^14.3.1",
     "@types/jasmine": "~4.0.0",
     "jasmine-core": "~4.1.0",
     "karma": "~6.3.0",
